fix: error on formatting empty list
The formatter threw an error in any files with an empty list following another parentheses. This is because `not elem` is true when elem is an empty List, causing the loop to break and for it to never get appended to the expression. Then, the error got raised, because `elem` was not a closing parentheses.